set modal eliminar

+++ public_html/js/repositorio.js
@@ -61,44 +61,109 @@ $(document).ready(function(){
 
     function eliminarUsuario(id){
 
-        //alert("entro");
+        
+        //var token = getToken($("#floatingInputCorreo").val(),$("#floatingInputPassword").val());
 
-        var url = "/usuarios/" + id;
-        //var parametros ={
-        //};
+        //************** */
+        
+        var token = null;
+
+        var url = "/auth/login";
+        var parametros = {
+            email: $("#floatingInputCorreo").val(),
+            password: $("#floatingInputPassword").val()
+        }
+
+        //console.log(parametros)
 
         $.ajax({
-            type: 'DELETE',
+            type: 'POST',
             url: url,
             //contentType: 'application/json',
-            //headers: {'Content-Type': 'application/x-www-form-urlencoded'},
-            
-            headers: {'x-access-token': 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NDhiZjFkNzNiNWQ1NDcwNTliYzlkNjUiLCJlbWFpbCI6ImdkemliQGxpdmUuY29tLm14IiwiaWF0IjoxNjg2ODkzMDk4LCJleHAiOjE2ODY4OTQ4OTh9.3JfcBbVI5a345_2GZk0DPkLXKSGAl48FCVJharnwR_Y'},
+            headers: {'Content-Type': 'application/x-www-form-urlencoded'},
             contentType: 'application/x-www-form-urlencoded; charset=utf-8',
             dataType: 'json',
             //data: JSON.stringify(parametros),
-            //data: parametros,    
+            data: parametros,    
         }).done(function(data) {
             
             console.log('SUCCESS');
-
-            var msj = ` <div class="alert alert-success" role="alert">
-                                <strong>¡Usuario eliminado exitosamente!<strong>
-                            </div>  `;
-
-            $("#respuestaModalE").html(msj)
+            console.log(data.token);
+            
+            token = data.token
+           //return data;
 
         }).fail(function (msg) {
 
-            var msj = ` <div class="alert alert-danger" role="alert">
-                            <strong>¡Ocurrio un error al eliminar usuario!<strong>
-                        </div>  `;
-
-            $("#respuestaModalE").html(msj)
+            return null;
 
         }).always(function (msg) {
+            console.log(msg)
+            console.log(msg.token.toString())
             console.log('ALWAYS');
+
+            token = msg.token
+            //return msg;
         });
+        
+
+
+        //********************** */
+
+        setTimeout(function(){
+
+                console.log("respuesta A: " + token)
+
+                if( !token ){
+                    var msj = ` <div class="alert alert-danger" role="alert">
+                                    <strong>¡Ocurrio un error al generar Token!<strong>
+                                </div>  `;
+
+                    $("#respuestaModalE").html(msj)
+
+                    return false;
+                }
+
+                console.log("respuesta D: " + token)
+
+                var url = "/usuarios/" + id;
+                //var parametros ={
+                //};
+
+                $.ajax({
+                    type: 'DELETE',
+                    url: url,
+                    //contentType: 'application/json',
+                    //headers: {'Content-Type': 'application/x-www-form-urlencoded'},
+                    
+                    headers: {'x-access-token': token},
+                    contentType: 'application/x-www-form-urlencoded; charset=utf-8',
+                    dataType: 'json',
+                    //data: JSON.stringify(parametros),
+                    //data: parametros,    
+                }).done(function(data) {
+                    
+                    console.log('SUCCESS');
+
+                    var msj = ` <div class="alert alert-success" role="alert">
+                                        <strong>¡Usuario eliminado exitosamente!<strong>
+                                    </div>  `;
+
+                    $("#respuestaModalE").html(msj)
+
+                }).fail(function (msg) {
+
+                    var msj = ` <div class="alert alert-danger" role="alert">
+                                    <strong>¡Ocurrio un error al eliminar usuario!<strong>
+                                </div>  `;
+
+                    $("#respuestaModalE").html(msj)
+
+                }).always(function (msg) {
+                    console.log('ALWAYS');
+                });
+
+        },5000);
 
     }
 
@@ -165,10 +230,59 @@ $(document).ready(function(){
         $.get("/usuarios",function(data){
 
             //alert(data[0].nombre);
-            console.log(data[0]._id);
+            //console.log(data[0]._id);
 
             $("#respuesta").html(data)
 
+        });
+
+    }
+
+//---------------------------------------------
+
+    function getToken(email,password){
+
+        var url = "/auth/login";
+        var parametros = {
+            email: email,
+            password: password
+        }
+
+        //console.log(parametros)
+
+        $.ajax({
+            type: 'POST',
+            url: url,
+            //contentType: 'application/json',
+            headers: {'Content-Type': 'application/x-www-form-urlencoded'},
+            contentType: 'application/x-www-form-urlencoded; charset=utf-8',
+            dataType: 'json',
+            //data: JSON.stringify(parametros),
+            data: parametros,    
+        }).done(function(data) {
+            
+            console.log('SUCCESS');
+            console.log(data.token);
+            /*
+
+            setTimeout(function(){
+                location.href ='index.html';
+            },3000);
+            */
+           return data.token
+           //return data;
+
+        }).fail(function (msg) {
+
+            return null;
+
+        }).always(function (msg) {
+            console.log(msg)
+            console.log(msg.token.toString())
+            console.log('ALWAYS');
+
+            return msg.token
+            //return msg;
         });
 
     }
